requirements.txt
alembic==1.8.1
coloredlogs==15.0.1
ffmpeg==1.4
ffmpeg_python==0.2.0
Flask==2.2.2
Flask_Bcrypt==1.0.1
Flask_Cors==3.0.10
Flask_Login==0.6.2
Flask_Mail==0.9.1
Flask_Migrate==3.1.0
Flask_Session==0.4.0
Flask_SQLAlchemy==2.5.1
psutil==5.9.2
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.41
tmdbsimple==2.9.1
